fixed problems with code for now; warning: code only works on 64-bit processors, so never give true to $forceThirtyTwoBitMath until a PHP integer replacement is written

+++ src/Order0Model.php
@@ -52,7 +52,7 @@ class Order0Model {
 			$total = 0;
 			
 			for($currentSymbol = 1; $currentSymbol < self::NUMBER_OF_SYMBOLS - 1; $currentSymbol++) {
-				$total += ((($this->frequencies[$currentSymbol] - $this->frequencies[$currentSymbol - 1])/2) + 1);
+				$total += (integer) ((($this->frequencies[$currentSymbol] - $this->frequencies[$currentSymbol - 1])/2) + 1);
 				$this->frequencies[$currentSymbol] = $this->frequencies[$currentSymbol - 1] + ((($this->frequencies[$currentSymbol] - $this->frequencies[$currentSymbol - 1])/2) + 1);
 			}
